# Management-student/src/tsoa/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"GioiTinh": {
				"enum": [
					"Nam",
					"Nu",
					"Khac"
				],
				"type": "string"
			},
			"TrangThaiHoatDong": {
				"enum": [
					"Danghoc",
					"TamNgunghoc",
					"KhongConHoc"
				],
				"type": "string"
			},
			"VaiTroNguoiDung": {
				"enum": [
					"SinhVien",
					"GiangVien",
					"Admin"
				],
				"type": "string"
			},
...
			"UpdateSinhVien": {
				"properties": {
					"MasoSinhVien": {
						"type": "string"
					},
					"HoVaTenSinhVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"GioiTinhHocSinh": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"NoiSinh": {
						"type": "string"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"KhoaHoc": {
						"type": "number",
						"format": "double"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDong"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double"
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"MasoSinhVien",
					"HoVaTenSinhVien",
					"NgaySinh",
					"GioiTinhHocSinh",
					"NoiSinh",
					"ChuyenNghanh",
					"KhoaHoc",
					"TrangThai",
					"Email",
					"UserName",
					"Password",
					"SoLamDangNhapThatBai",
					"KhongChoDangNhapToi"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TrangThaiHoatDongGiangVien": {
				"enum": [
					"DangDay",
					"TamNgungDay",
					"KhongConDay"
				],
				"type": "string"
			},
			"CreateGiangVien": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"TenGiangVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"NoiSinh": {
						"type": "string"
					},
					"GioiTinhGiangVien": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDongGiangVien"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"default": null
					}
				},
				"required": [
					"MaSoGiangVien",
					"TenGiangVien",
					"NgaySinh",
					"NoiSinh",
					"GioiTinhGiangVien",
					"ChuyenNghanh",
					"TrangThai",
					"Email",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateGiangVien": {
				"properties": {
					"MaSoGiangVien": {
						"type": "string"
					},
					"TenGiangVien": {
						"type": "string"
					},
					"NgaySinh": {
						"type": "string",
						"format": "date-time"
					},
					"NoiSinh": {
						"type": "string"
					},
					"GioiTinhGiangVien": {
						"$ref": "#/components/schemas/GioiTinh"
					},
					"ChuyenNghanh": {
						"type": "string"
					},
					"TrangThai": {
						"$ref": "#/components/schemas/TrangThaiHoatDongGiangVien"
					},
					"VaiTro": {
						"$ref": "#/components/schemas/VaiTroNguoiDung"
					},
					"Email": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"SoLamDangNhapThatBai": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"KhongChoDangNhapToi": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"default": null
					}
				},
				"required": [
					"MaSoGiangVien",
					"TenGiangVien",
					"NgaySinh",
					"NoiSinh",
					"GioiTinhGiangVien",
					"ChuyenNghanh",
					"TrangThai",
					"Email",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DangNhapDto": {
				"properties": {
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrangThaiDangKi": {
				"enum": [
					"DaDangKi",
					"DaHuy"
				],
				"type": "string"
			},
			"CreateDangKihocPhan": {
				"properties": {
					"MasoSinhVien": {
						"type": "string"
					},
					"MaMonHoc": {
						"type": "string"
					},
					"MaLopHocPhan": {
						"type": "string"
					},
					"TrangThaiDangKi": {
						"$ref": "#/components/schemas/TrangThaiDangKi"
					}
				},
				"required": [
					"MasoSinhVien",
					"MaMonHoc",
					"MaLopHocPhan"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "management-student",
		"version": "0.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/API/v1/SinhVien/GetOneSinhVien/{MasoSinhVien}": {
			"get": {
				"operationId": "GetOneSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin sinh viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/SinhVien/GetAllSinhVien": {
			"get": {
				"operationId": "GetAllSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin tất cả sinh viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/SinhVien/UpdateOneSinhVien/{MasoSinhVien}": {
			"put": {
				"operationId": "UpdateOneSinhVien",
				"responses": {
					"200": {
						"description": "Cập nhập thông tin sinh viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSinhVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/SinhVien/DeleteOneSinhVien/{MasoSinhVien}": {
			"delete": {
				"operationId": "DeleteOneSinhVien",
				"responses": {
					"200": {
						"description": "Xoá sinh viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SinhVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/API/v1/GiangVien/CreateGiangVien": {
			"post": {
				"operationId": "PostGiangVien",
				"responses": {
					"201": {
						"description": "Insert giảng viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateGiangVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/GiangVien/GetOneGiangVien/{MaSoGiangVien}": {
			"get": {
				"operationId": "GetOneGiangVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin giảng viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/GiangVien/GetAllGiangVien": {
			"get": {
				"operationId": "GetAllSinhVien",
				"responses": {
					"200": {
						"description": "Lấy thông tin tất cả giảng viên thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/API/v1/GiangVien/UpdateOneGiangVien/{MaSoGiangVien}": {
			"put": {
				"operationId": "UpdateOneSinhVien",
				"responses": {
					"200": {
						"description": "Cập nhập thông tin giảng viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateGiangVien"
							}
						}
					}
				}
			}
		},
		"/API/v1/GiangVien/DeleteOneGiangVien/{MaSoGiangVien}": {
			"delete": {
				"operationId": "DeleteOneSinhVien",
				"responses": {
					"200": {
						"description": "Xoá giảng viên thành công.",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"GiangVien"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MaSoGiangVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/DangNhap/Login/{UserName}/{Password}": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Đăng nhập thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DangNhap"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DangNhapDto"
							}
						}
					}
				}
			}
		},
		"/API/v1/DangNhap/LoginAgain": {
			"post": {
				"operationId": "LoginAgain",
				"responses": {
					"200": {
						"description": "Đăng nhập lại thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DangNhap"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"RefreshToken": {
										"type": "string"
									}
								},
								"required": [
									"RefreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/API/v1/DangNhap/Logout": {
			"post": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "Logout thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DangNhap"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"RefreshToken": {
										"type": "string"
									}
								},
								"required": [
									"RefreshToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/API/v1/DangKiHocPhan/CreateDangKiHocPhan": {
			"post": {
				"operationId": "CreateDangKiHocPhan",
				"responses": {
					"201": {
						"description": "Tạo thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DangKiHocPhan"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDangKihocPhan"
							}
						}
					}
				}
			}
		},
		"/API/v1/DangKiHocPhan/TimKiemLopHocPhanDaDangKi/{MasoSinhVien}/{MaMonHoc}": {
			"get": {
				"operationId": "TImKiemLopHocPhanDaDangKi",
				"responses": {
					"200": {
						"description": "Tìm kiếm thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"DangKiHocPhan"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/DangKiHocPhan/DoiLophocPhan/{MasoSinhVien}/{MaMonHoc}/{MaLopHocPhan}": {
			"put": {
				"operationId": "DoiLophocPhan",
				"responses": {
					"200": {
						"description": "Đã đổi học phần thành công",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"DangKiHocPhan"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaLopHocPhan",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/API/v1/DangKiHocPhan/HuyDangKiHocPhan/{MasoSinhVien}/{MaMonHoc}": {
			"delete": {
				"operationId": "HuyDangKiHocPhan",
				"responses": {
					"200": {
						"description": "Huỷ thành công",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"DangKiHocPhan"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "MasoSinhVien",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "MaMonHoc",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}